requirements.txt
fastapi
uvicorn
google-generativeai
psycopg2-binary
sqlalchemy
pgvector
typing
langgraph
langchain
langchain-google-genai
python-dotenv
redshift-connector
PyPDF2
python-multipart
pydantic
langchain-core
langchain-qdrant
langchain-postgres
qdrant-client
pandas
sentence-transformers
dotenv

